requirements.txt
pyrogram
tgcrypto
python-dotenv
asyncio
# aiohttp
uvloop
# pytz